Fix leftover uint16_t/default-value gaps from event-logging sentinel narrowing
Commit dbccfd1 narrowed infector_symptom_id (and related event-logging
fields) to uint8_t and introduced kUnknownEncounterTypeId/kNoSymptomId
sentinels, but two call sites weren't updated to match:

- interaction_manager_partial_presence.cpp: unpacking MPI-packed
  candidate data cast infector_symptom_id to uint16_t, then relied on an
  implicit second narrowing to fit PendingInfection's now-uint8_t field.
  Cast directly to uint8_t instead, matching the field type and the
  explicit style used at the sibling call site in
  interaction_manager_venue.cpp.

- infection_seed.cpp: logInfection call for seed infections omitted the
  encounter-type and symptom-id sentinel arguments added by dbccfd1
  (no infector exists for seeded infections), leaving them at
  compiler/default-initialised values instead of the intended
  kUnknownEncounterTypeId/kNoSymptomId sentinels.

+++ src/interaction/interaction_manager_partial_presence.cpp
@@ -460,7 +460,7 @@ int InteractionManager::resolvePartialPresenceInfections(
     c.venue_id = static_cast<VenueId>(packed[i + 1]);
     c.infector_id = static_cast<PersonId>(packed[i + 2]);
     c.venue_type_id = static_cast<uint8_t>(packed[i + 3]);
-    c.infector_symptom_id = static_cast<uint16_t>(packed[i + 4]);
+    c.infector_symptom_id = static_cast<uint8_t>(packed[i + 4]);
     c.transmission_mode_index = static_cast<uint8_t>(packed[i + 5]);
     c.infection_time = current_time;
     c.encounter_type_id = 255;
